% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ace2fastq-package.R
\docType{package}
\name{ace2fastq-package}
\alias{ace2fastq}
\alias{ace2fastq-package}
\title{ace2fastq: ACE File To FASTQ Converter}
\description{
Converts a sequence in .ace file format to .fastq format.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/c5sire/ace2fastq}
  \item Report bugs at \url{https://github.com/c5sire/ace2fastq/issues}
}

}
\author{
\strong{Maintainer}: Reinhard Simon \email{rsimon64@cgiar.org} (https://orcid.org/0000-0002-4608-9077)

}
\keyword{internal}
