% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ace2fastq-package.R
\docType{package}
\name{ace2fastq}
\alias{ace2fastq}
\alias{ace2fastq-package}
\title{ace2fastq}
\description{
Converts a sequence in .ace file format to .fastq format.
}
\section{ace2fastq functioniality}{


See: \code{\link{ace_to_fastq}}
}

