% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ace2fastq.R
\name{ace_to_fastq}
\alias{ace_to_fastq}
\title{ace_to_fastq}
\usage{
ace_to_fastq(filename, target_dir = dirname(filename), name2id = TRUE)
}
\arguments{
\item{filename}{.ace file}

\item{target_dir}{target directory or stdout}

\item{name2id}{use the file name as primary id or not. Default is TRUE.}
}
\value{
list
}
\description{
Converts one or more contig sequences in .ace file format to .fastq format.
The parameter target_dir has a special value 'stdout' which will just return
the contigs as a list.
}
\examples{

  library(ace2fastq)
  filename <- system.file("sampledat/1.seq.ace", package = "ace2fastq")
 
  fileout <- ace_to_fastq(filename, target_dir = tempdir())

}
\author{
Reinhard Simon
}
