% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.R
\name{print.ace2fastq}
\alias{print.ace2fastq}
\title{print.ace2fastq}
\usage{
\method{print}{ace2fastq}(x, ...)
}
\arguments{
\item{x}{an ace2fastq object}

\item{...}{other print parameters}
}
\description{
print method for an object of class ace2fastq
}
\author{
Reinhard Simon
}
\concept{ace2fq_utils}
