% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ace2fastq.R
\name{ace_to_fastq}
\alias{ace_to_fastq}
\title{ace_to_fastq}
\usage{
ace_to_fastq(filename, target_dir = dirname(filename), name2id = TRUE)
}
\arguments{
\item{filename}{.ace file}

\item{target_dir}{target directory}

\item{name2id}{use the file name as primary id or not. Default is TRUE.}
}
\value{
target file name
}
\description{
ace_to_fastq
}
\examples{
if(interactive()) {
  library(ace2fq)
  filename <- system.file("sampledat/1.seq.ace", package = "ace2fq")
  ace_to_fastq(filename)
}


}
\author{
Reinhard Simon
}
